Implemented /stats microservice.

+++ src/main/java/com/squidev/mutants_identifier/dtos/DTOStats.java
@@ -5,34 +5,35 @@ package com.squidev.mutants_identifier.dtos;
  */
 
 public class DTOStats {
-    int count_mutant_dna;
-    int count_human_dna;
-    int ratio;
+    //This structure will be converted automatically to JSON: {“count_mutant_dna”:40, “count_human_dna”:100: “ratio”:0.4}
+    long count_mutant_dna;
+    long count_human_dna;
+    double ratio;
 
     public DTOStats() {
     }
 
-    public int getCount_mutant_dna() {
+    public long getCount_mutant_dna() {
         return count_mutant_dna;
     }
 
-    public void setCount_mutant_dna(int count_mutant_dna) {
+    public void setCount_mutant_dna(long count_mutant_dna) {
         this.count_mutant_dna = count_mutant_dna;
     }
 
-    public int getCount_human_dna() {
+    public long getCount_human_dna() {
         return count_human_dna;
     }
 
-    public void setCount_human_dna(int count_human_dna) {
+    public void setCount_human_dna(long count_human_dna) {
         this.count_human_dna = count_human_dna;
     }
 
-    public int getRatio() {
+    public double getRatio() {
         return ratio;
     }
 
-    public void setRatio(int ratio) {
+    public void setRatio(double ratio) {
         this.ratio = ratio;
     }   
 }
